+++ ChangeLog.md
@@ -33,6 +33,12 @@
 
 
 
+#### todos
+
+- [ ] `bin.Command` 中 `App` 属性无时效性，需要在继承的命令中调用 `bin.GetApp()` 实现，需要优化 `runAppRouter()`
+
+
+
 
 #### alpha1/191025